Cleaning up code.... Uh, maybe just scrap the repo I copied from. Bad code practices, most code can be simplified and reduced by half, made more clear.

+++ noisedive_flask/routes/deleteUser.py
@@ -1,55 +1,34 @@
-from helpers import (
+from noisedive_flask.helpers import (
     session,
     sqlite3,
     message,
     redirect,
     Blueprint,
+    query,
+    commit_to_db
 )
 
 deleteUserBlueprint = Blueprint("deleteUser", __name__)
 
-
 @deleteUserBlueprint.route("/deleteuser/<userName>/redirect=<direct>")
 def deleteUser(userName, direct):
     direct = direct.replace("&", "/")
-    match "userName" in session:
-        case True:
-            connection = sqlite3.connect("db/users.db")
-            cursor = connection.cursor()
-            cursor.execute(
-                f'select * from users where lower(userName) = "{userName.lower()}"'
-            )
-            user = cursor.fetchone()
-            cursor.execute(
-                f'select role from users where userName = "{session["userName"]}"'
-            )
-            perpetrator = cursor.fetchone()
-            match user:
-                case None:
-                    message("1", f'USER: "{userName}" NOT FOUND')
-                    return redirect(f"/{direct}")
-                case _:
-                    match user[1] == session["userName"] or perpetrator[0] == "admin":
-                        case True:
-                            cursor.execute(
-                                f'delete from users where lower(userName) = "{userName}"'
-                            )
-                            cursor.execute(f"update sqlite_sequence set seq = seq-1")
-                            connection.commit()
-                            message("2", f'USER: "{userName}" DELETED')
-                            match perpetrator[0] == "admin":
-                                case True:
-                                    return redirect(f"{direct}")
-                                case False:
-                                    session.clear()
-                                    return redirect(f"{direct}")
-                        case False:
-                            message(
-                                "1",
-                                f'USER: "{user[1]}" NOT DELETED YOU ARE NOT {user[1]}',
-                            )
-                            return redirect(f"{direct}")
+    if "userName" in session:
+        user = query(f'select * from users where lower(userName) = "{userName.lower()}"', fetchone=True)
+        perpetrator = query(f'select role from users where userName = "{session["userName"]}"', fetchone=True)
+        if user is None:
+            return redirect(f"/{direct}")
+        else:
+            if user[1] == session["userName"] or perpetrator[0] == "admin":
+                commit_to_db(
+                    f'delete from users where lower(userName) = "{userName}"'
+                )
+                commit_to_db(f"update sqlite_sequence set seq = seq-1")
 
-        case False:
-            message("1", f"USER NEEDS TO LOGIN FOR DELETE USER: {userName}")
-            return redirect(f"/login/redirect=&deleteuser&{userName}&redirect=index")
+                if perpetrator[0] != "admin":
+                    session.clear()
+
+            return redirect(f"{direct}")    
+
+    else:
+        return redirect(f"/login/redirect=&deleteuser&{userName}&redirect=index")
